Adjust removal

+++ schema.js
@@ -137,7 +137,7 @@ module.exports = class Schema {
     let retObject = {};
     for (const key in this._fieldDefs) {
       // TODO: Build in required properties.
-      if (object[key] === undefined) {
+      if (object.hasOwnProperty(key) && object[key] === undefined) {
         retObject[key] = Firestore.FieldValue.delete();
       } else if (object.hasOwnProperty(key) && typeof object[key] !== 'undefined') {
         try {
